hv:clear up the usage of printf data struct
The printf related functions have been called by other various
functions, e.g. get_ptdev_info(), get_ioapic_info, etc. The patch
is used to clear up the usage.

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/assign.c
@@ -798,11 +798,11 @@ static void get_entry_info(struct ptdev_remapping_info *entry, char *type,
 	}
 }
 
-void get_ptdev_info(char *str_arg, int str_max)
+void get_ptdev_info(char *str_arg, size_t str_max)
 {
 	char *str = str_arg;
 	struct ptdev_remapping_info *entry;
-	int len, size = str_max;
+	size_t len, size = str_max;
 	uint32_t irq, vector;
 	char type[16];
 	uint64_t dest;
