+++ hypervisor/include/arch/x86/guest/vioapic.h
@@ -62,7 +62,7 @@ int	vioapic_mmio_access_handler(struct vcpu *vcpu,
 	struct io_request *io_req);
 
 #ifdef HV_DEBUG
-void get_vioapic_info(char *str_arg, int str_max, uint16_t vmid);
+void get_vioapic_info(char *str_arg, size_t str_max, uint16_t vmid);
 #endif /* HV_DEBUG */
 
 #endif
